refactor: rename `@modules/boxSources`
