+++ src/components/MagicBox/index.tsx
@@ -3,6 +3,7 @@ import CustomizedSnackbar from '@components/Snackbar';
 import copyTextToClipboard from '@functions/clipboard';
 import { trim } from '@functions/helper';
 import { Box, BoxOptions } from '@modules/Box';
+import { BoxSource } from '@modules/BoxSource';
 import {
   Base64DecodeBoxSource,
   Base64EncodeBoxSource,
@@ -21,8 +22,7 @@ import {
   TimestampBoxSource,
   URLDecodeBoxSource,
   UuidBoxSource,
-} from '@modules/boxes';
-import { BoxSource } from '@modules/BoxSource';
+} from '@modules/boxSources';
 import React, { useEffect, useState } from 'react';
 
 const defaultBoxSources: BoxSource[] = [
