+++ src/modules/boxSources/GenerateQRCodeBoxSource.ts
