+++ src/modules/boxSources/__test__/UuidBoxSource.test.ts
@@ -1,7 +1,7 @@
 import { DefaultBoxTemplate } from '@components/BoxTemplate';
 import { expect } from '@jest/globals';
 
-import { UuidBoxSource } from '../UuidBox';
+import { UuidBoxSource } from '../UuidBoxSource';
 
 describe('UuidBoxSource', () => {
   const validUUIDv4Regex = /^[0-9a-f]{8}-[0-9a-f]{4}-4[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$/i;
