--- src/modules/boxes/index.ts
@@ -1,16 +1,0 @@
-export * from './Base64Box';
-export * from './CronExpressionBox';
-export * from './DateCalculateBox';
-export * from './GenerateQRCodeBox';
-export * from './JWTBox';
-export * from './K8sSecretBox';
-export * from './MathExpressionBox';
-export * from './MyIPBox';
-export * from './NowBox';
-export * from './PrettyJSONBox';
-export * from './RandomIntegerBox';
-export * from './ShortenURLBox';
-export * from './TimeFormatBox';
-export * from './TimestampBox';
-export * from './URLDecodeBox';
-export * from './UuidBox';
